move useProxiedPath to hooks folder
